[GH-#349] Moved function refund_sender
